requirements.txt
cryptography
pillow
pyzbar
